backend/requirements.txt
# ─────────────────────────────────────────────
# Web framework & server
# ─────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
websockets>=13.0

# ─────────────────────────────────────────────
# Data validation & settings
# ─────────────────────────────────────────────
pydantic[email]>=2.9.0
pydantic-settings>=2.5.0

# ─────────────────────────────────────────────
# Database
# ─────────────────────────────────────────────
sqlalchemy>=2.0.30
psycopg[binary]>=3.2.0
alembic>=1.13.0

# ─────────────────────────────────────────────
# Auth
# ─────────────────────────────────────────────
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# ─────────────────────────────────────────────
# AI / Agent
# ─────────────────────────────────────────────
langgraph>=0.2.50
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
openai>=1.50.0
tiktoken>=0.7.0

# ─────────────────────────────────────────────
# Vector store (embedded)
# ─────────────────────────────────────────────
chromadb>=0.5.0

# ─────────────────────────────────────────────
# Mock data
# ─────────────────────────────────────────────
faker>=30.0.0
